Make the port passed to qbx's query_q_server() an unsigned short, which eliminates a warning.
git-svn-id: svn://svn.code.sf.net/p/bitchx/code/trunk@323 13b04d17-f746-0410-82c6-800466cd88b0

+++ dll/qbx/qbx.c
@@ -181,7 +181,7 @@ void q_timer(int fd) {
 	querying = 0;
 }
 
-void query_q_server(char *server, int port, int game) {
+void query_q_server(char *server, unsigned short port, int game) {
 	char buffer[16];
 
 	struct sockaddr_in qsock;
@@ -227,7 +227,7 @@ void query_q_server(char *server, int port, int game) {
 
 int pub_proc(char *which, char *str, char **unused) {
 	char *loc, *nick, *chan, *cmd, *serv;
-	int port = 0;
+	unsigned short port = 0;
 
 	if(qbx_on == 0) return 1;
 	loc = LOCAL_COPY(str);
